Move movie interfaces to models
