+++ pkg/scene/import.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/stashapp/stash/pkg/models"
 	"github.com/stashapp/stash/pkg/models/jsonschema"
-	"github.com/stashapp/stash/pkg/movie"
 	"github.com/stashapp/stash/pkg/performer"
 	"github.com/stashapp/stash/pkg/sliceutil/stringslice"
 	"github.com/stashapp/stash/pkg/tag"
@@ -25,7 +24,7 @@ type Importer struct {
 	StudioWriter        models.StudioFinderCreator
 	GalleryFinder       models.GalleryFinder
 	PerformerWriter     performer.NameFinderCreator
-	MovieWriter         movie.NameFinderCreator
+	MovieWriter         models.MovieFinderCreator
 	TagWriter           tag.NameFinderCreator
 	Input               jsonschema.Scene
 	MissingRefBehaviour models.ImportMissingRefEnum
